Make dependency installation more reliable
Checks whether build tools such as `gcc`, `autoconf` and `make` exist
in the system at which these dependencies are going to be installed (Issue #11).

Also more information is printed during installation.

Color will be displayed in printed log.

Bug in which the installation script didn't stop even though one of the dependency failed to be installed,
was fixed by using `$PIPESTATUS` instead of `$?` to capture the exit status.